Refactor tests and add more tests
- Extract fixtures to conftest.py
- Add test and facit for Formula.evaluate function
